Quarantine legacy graph extractors and redact LifeTrack trace sinks
Move LifeTrackFactExtractor and ChatEntityGraphProjector into addons/graph_workflow
with host export stubs. Add invocation epoch cancellation for mid-ingest revoke,
redact sensitive tool parameters in orchestrator traces, and add lifecycle/trace
sink tests plus architecture gate cleanup.
